ai_server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
bcrypt==5.0.0
build==1.4.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.2
colorama==0.4.6
comm==0.2.3
cryptography==46.0.7
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
durationpy==0.10
executing==2.2.1
fastapi==0.135.3
filelock==3.28.0
flatbuffers==25.12.19
fsspec==2026.3.0
google-auth==2.49.2
google-genai==1.73.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==7.1.0
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kubernetes==35.0.0
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==2.4.3
oauthlib==3.3.1
onnxruntime==1.24.4
opencv-python==4.13.0.92
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.8
overrides==7.7.0
packaging==26.0
parso==0.8.6
platformdirs==4.9.4
prompt_toolkit==3.0.52
protobuf==6.33.6
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
Pygments==2.19.2
pypdf==6.10.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
zipp==3.23.1
